Ensure Fragment subclasses can be initiated.
From the Fragment documentation: Every fragment must have an empty
constructor, so it can be instantiated when restoring its activity's
state. It is strongly recommended that subclasses do not have other
constructors with parameters, since these constructors will not be
called when the fragment is re-instantiated; instead, arguments can be
supplied by the caller with setArguments(Bundle) and later retrieved by
the Fragment with getArguments().